refactor: add PaginatedResult<T> in Application.Shared and extract IssuedInvoiceSyncStats to its own file
