+++ backend/src/Anela.Heblo.Application/Shared/PaginatedResult.cs
@@ -1,0 +1,16 @@
+namespace Anela.Heblo.Application.Shared;
+
+/// <summary>
+/// Paginated result container
+/// </summary>
+/// <typeparam name="T">Type of items</typeparam>
+public class PaginatedResult<T>
+{
+    public IEnumerable<T> Items { get; set; } = new List<T>();
+    public int TotalCount { get; set; }
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
+    public int TotalPages => (int)Math.Ceiling((double)TotalCount / PageSize);
+    public bool HasNextPage => PageNumber < TotalPages;
+    public bool HasPreviousPage => PageNumber > 1;
+}
